New config option that determines whether you can only lose max hearts when killed by a MOB. If enabled, it overrides the other option to lose max hearts only when killed by a player.
Major bug fixed that has been in the game for quite a while. If you damaged a player/mob and it died by fire, you would not get hearts. This fix makes it so whenever you damage a player/mob, if they die within the next 100 ticks/5 seconds it will count as you killing them.

A little bit of code rewrite to make things more consistent.

There is a known incompatibility with Regeneration Mod at the moment. Whenever you kill a player to gain hearts and they regenerate, it'll still count as a kill in this mod giving you hearts. I won't be able to fix this for now because my tiny brain cannot handle reading its code.